fixed some logic errors and adjusted some types.
